Fix for missing data in company/moa tab in Trialzilla.

+++ preindex_trial.php
@@ -210,17 +210,25 @@ function tindex($sourceid,$cat,$productz=NULL,$up_id=NULL,$cid=NULL,$productID=N
 					else
 					{
 						//delete existing product/area indexes
-						$qry='DELETE from '. $table .' where `'. $field . '` = "'. $productID . '"';
-						if(!mysql_query($qry))
+						
+						
+						$query = 'SELECT LI_id from '. 'entities' .' where `id`="' . $productID .'" limit 1' ;
+						$resu = mysql_query($query);
+						$row=mysql_fetch_array($resu);
+						if(!empty($row['LI_id'])) //delete only if they are not mesh related indexes
 						{
-							$log='Could not delete existing product indexes. Query='.$qry.' Error:' . mysql_error();
-							$logger->fatal($log);
-							$query = 'update update_status_fullhistory set 
-							er_message="' . $log . '" where update_id= "'. $up_id .'" limit 1' ; 
-							mysql_query($query);
-							mysql_query('ROLLBACK');
-							echo $log;
-							return false;
+							$qry='DELETE from '. $table .' where `'. $field . '` = "'. $productID . '"';
+							if(!mysql_query($qry))
+							{
+								$log='Could not delete existing product indexes. Query='.$qry.' Error:' . mysql_error();
+								$logger->fatal($log);
+								$query = 'update update_status_fullhistory set 
+								er_message="' . $log . '" where update_id= "'. $up_id .'" limit 1' ; 
+								mysql_query($query);
+								mysql_query('ROLLBACK');
+								echo $log;
+								return false;
+							}
 						}
 //						$query = substr($mystring,0,$pos+6). '  ( ' . substr($mystring,$pos+6) . ' ) ';
 						$query = $mystring ;
@@ -616,9 +624,18 @@ function tindex($sourceid,$cat,$productz=NULL,$up_id=NULL,$cid=NULL,$productID=N
 			}
 		}
 	}
+	
+	
+	
 	elseif(isset($productID) and !empty($productID))
 	{
-	$qry='DELETE from '. $table .' where `'. $field . '` = "'. $productID . '"';
+	
+		$query = '	SELECT LI_id from '. 'entities' .' where `id`="' . $productID .'" limit 1' ;
+					$resu = mysql_query($query);
+					$row=mysql_fetch_array($resu);
+					if(!empty($row['LI_id'])) //delete only if they are not mesh related indexes
+					{
+						$qry='DELETE from '. $table .' where `'. $field . '` = "'. $productID . '"';
 						if(!mysql_query($qry))
 						{
 							$log='Could not delete existing product indexes. Query='.$qry.' Error:' . mysql_error();
@@ -630,6 +647,7 @@ function tindex($sourceid,$cat,$productz=NULL,$up_id=NULL,$cid=NULL,$productID=N
 							echo $log;
 							return false;
 						}
+					}
 	}
 }
